Replace hard-coded relative URL
This should be the last hard-coded HTTP URL, finishing the work on EDD-1261.
All other instances I could find have been removed during the course of
other work.

There remains a few more instances of hard-coding, related to WebSocket URLs. I
believe these can be excused for now, as Channels does not have the same
support for naming routes as Django overall does for naming paths, and the
WebSocket routes used by EDD are limited to two distinct paths, unlikely to
expand in the near-to-mid-future. For those links, `/ws/notify/` and
`/ws/load/` are de-facto names for the routes, and changing the code to remove
them does little to improve the code.