Fix blank modal in empty study
Fix for EDD-1340. When navigating to the Lines view of a Study, the function to
set up the page modal dialogs was tied to the loading of the Handsontable
widget. Since the table never loads when there is no line data, the modals are
not properly set up in an empty Study. Moving the modal setup to the callback
for page loading allows for all the correct setup on the modal, regardless of
whether the table loads.