Improve build process for edd-core Dockerfile
A handful of small tweaks to how the Dockerfile for the edd-core image is
building. The result of these tweaks are a final image that's about 40MB
smaller, and builds on my laptop around 25-30 seconds faster.

Adding an additional build argument, `REPO_SERVER`, for specifying which
container registry to use for base images. The default is set to "docker.io",
which is the same as just leaving it blank. This will let our automated builds
specify our local registry instead during builds, to avoid needing to pull or
push from the public "docker.io" container registry.

The build phases are now more strictly separating the TypeScript builds from
the Python code. There should no longer be a need to wait for TypeScript to
rebuild if only Python code is changing, which will save significant time;
TypeScript builds typically take over a minute. An additional build phase is
added to load in the current git commit hash, which was previously being done
along with the TypeScript build. The Python code is now loaded in directly,
instead of being passed along from the TypeScript build phase. This means that
Python-only changes can now be rebuilt in a matter of seconds.