Fix display of non-metabolites with compartment
The filter code was making assumptions that all non-metabolite measurements
would always have a "0" or "N/A" value set on compartments. This turns out to
not be the case. We still only distinguish between compartments for
metabolites, but now a measurement with a non-N/A compartment will properly
show up in graphs.