Adding simplified Strain Registry API
Including an updated and simplified API to interact with ICE, as part of the
edd.search app. Replacing integration tests using the old API with updated
tests using the new API; this is cutting down significantly on the number of
warnings emitted for hanging sockets after running the test suites.