Replace patched cleanse_setting with new API
Django prior to 3.2 used a function `django.views.debug.cleanse_setting()` to
replace any potentially sensitive parameters in settings with an obfuscated
value (a string of asterisks). In Django 3.2+, this function is replaced by a
SafeExceptionReporterFilter class, that handles cleansing settings, as well as
sensitive parameters in request META headers and post parameters, when
reporting values in errors.

EDD was using a monkey-patched version of `cleanse_setting`, that checked for
keys likely to contain URLs, then parse those URLs and replace any found
passwords with asterisks. Now that the original function is gone and refactored
into a class swappable in settings, moved the patched code into a sub-class of
the new class, and changed the base EDD settings to load our version of
the class.