Fix deleting of lines
Django as of 3.2 dis-allows doing QuerySet .delete() calls after calls to
.distinct(), because in some cases there are reports of .delete() after
.distinct() generating SQL like `DELETE FROM table;`. Deletion of lines is
making use of the ExportSelection class, that can take identifiers for studies,
lines, assays, and measurements, and create a selection that includes all
parent objects for the given identifiers, and all child objects. This ends up
using a call to .distinct() for each level. Getting around the limitation by
selecting the id values for lines, then doing a delete on lines having id in
that set.