Replace contrib JSONField with Django JSONField
Django now has bundled support for a JSONField on models, and the same field in
the django.contrib.postgres package is now deprecated. The contrib version will
be removed in Django 4.0. See:

https://docs.djangoproject.com/en/3.2/releases/3.1/#deprecated-jsonfield

The migration generated from the implementation switch is a SQL no-op, so
modifying the existing migrations instead of adding further no-op files to the
migration history.