Improve message for limiting values in plots
The message displayed when a plot has stopped adding additional values to
display on the Study Data view has some updated wording to make it clearer what
is happening. It's also made more prominent by using the warning alert class,
so it shows as a larger yellow warning banner instead of just a line of text.