Add spacing for export button in Lines view
On production deployments, the "Report a bug" issue collector from Jira was
getting rendered near the spot where the Export button is placed, partially
obscuring the button. Adding some additional spacing to account for the 24
pixel height of the issue collector button, and clarified the components of the
fudge factor used to calculate the sizing of the table and positioning of the
controls on the bottom of the table.