Change y-axis extent calculation
The previous calculation made assumptions about wanting a zero-bound y-axis,
which may not hold. It also forced the difference between max and min to be at
least one.

The new calculation does not force the y-axis to contain zero, and only forces
the difference between max and min to be one when the actual difference
is zero.

To demonstrate:

Range: 6.85 to 7.12
Old axis: 0 to 7.12
New axis: 6.85 to 7.12

Range: 0.021 to 0.054
Old axis: 0 to 1
New axis: 0.021 to 0.054

Range: 0 to 0
Old axis: 0 to 1
New axis: 0 to 1

Range: 42 to 42
Old axis: 0 to 42
New axis: 42 to 43