Normalize export button between views
The "Export Data" button available on the Description (Lines) view, and the
Data view, was inconsistent. Only the Description view had options on the
button, while the Data view only supported the CSV export. Adding in the menu
options available from the Description view version of the button to the Data
view version.