Add Redis-based error tracking for edd.describe
The edd.describe.broker.DescribeErrorReport class added here (lifted from edd.load) supports the immediate need to refactor email-based reporting of errors and warnings, and will also likely be useful as a basis for later, wider use of Redis similar to edd.load’s LoadRequest.  Also lifted the subset of relevant tests from edd.load.  Likely in the future that existing CombinatorialDescriptionInput, ExperimentDescriptionOptions, and this new DescribeErrorReport may be combined to further simplify code (with much wider scope of changes). Purposefully deferring that change until a later round of refactoring to limit the scope of code changes in this round.