Re-lock Pipfile to fix test launch
A regression in Python 3.8.10 is causing the loading of the faker module to
fail. The project has released a fix with faker version 8.1.3 that should not
trigger the upstream regression.

This is also updating several other dependencies, as I ran `pipenv lock` rather
than updating just the individual testing dependency. There are a few pending
updates to Django, Pillow, etc anyway, so this commit will resolve
those updates.