Use container static files as default fallback
Building the EDD image will bundle in static assets with the Docker image. When
running in development with volume-mounted code, there will often not be any
static assets saved in the working development directory. This is causing a
fresh clone and startup to fail, as EDD will try to read static assets from the
volume-mounted directory, and nothing will be there in a fresh clone. Setting
the bundled version as an additional location with STATICFILES_DIRS will allow
those files to be found, and for startup to complete.