Include localhost in default ALLOWED_HOSTS setting
The previous default of an empty list for ALLOWED_HOSTS was causing health
checks on the http and websocket containers to fail, because Django returns a
500 response for a request not on an allowed host. Adding this change allows
EDD to work correctly from a fresh clone as described in the documentation.
Real deployments should still initialize code in the local settings module and
assign real host strings to the ALLOWED_HOSTS setting.