Remove references to upload size limits
EDD no longer is storing files on local disks, and is instead using S3 buckets.
This, combined with using Traefik for a proxy, eliminates the technical
requirements to have upload file size limits and defaults. A few places in the
code were still referencing these limits, and are now removed.