Fix loading of ASGI application
The upgraded Channels library makes some changes to how ASGI applications are
loaded. Updating our code to more closely match what the 3.x documentation
shows. Wrapping the imports of EDD custom code and AuthMiddlewareStack in a
function, as the default example shown in the documentation will have imports
re-organized in a way that will break the application.