Fix parsing of files with BOM
Excel will sometimes save CSV files with a Unicode Byte Order Marker, using the
zero-width non-breaking space character, which shows up in Python as "\ufeff".
When EDD gets a file like this, the BOM is not stripped from the string,
causing a failure to match to expected column headers. This adds a test for the
situation where there's a BOM in a file, and changes the code to pass the test.