Fix AutocompleteRenderer for Assay experimenter
The form field renderer for an assay experimenter was using an outdated
refactoring, where it was expecting to get an EDDContact in the render
callback. The final refactoring actually sends an argument of the AssayRecord.
The code is now properly dereferencing the callback parameter and providing the
correct information for rendering form elements.