Add edit_uri key to mkdocs.yml
The default MkDocs implementation assumes that all GitHub repos use "master" as
the default branch. We use "trunk" as the default branch, and the "master"
branch no longer exists. This assumption causes the link for "Edit in GitHub"
to return a 404 page, as the link goes to a branch that does not exist.

Changing the edit_uri key to explicitly use a path to edit in "trunk" fixes the
problem. See [the edit_uri docs][1], specifically the note section that says:

    For example, for a GitHub- or GitLab-hosted repository, the `edit_uri`
    would be automatically set as `edit/master/docs/` (Note the `edit` path
    and `master` branch).

    …

    To use a different URI than the default (for example a different branch),
    simply set the `edit_uri` to your desired string.

[1]: https://www.mkdocs.org/user-guide/configuration/#edit_uri