Refactor merging of replicates to own class
Moving the logic for scanning through an array of LineRecord objects to merge
together those having the same replicate UUID, out of the replicates function
of the Access object, and into a dedicated class for merging replicates. The
method was out-of-place in the Access object, as all the others were only a
handful of lines, and not 40-ish.