Refactor how line data is loaded into Handsontable
Remove the explicit method on Access to get replicates; it's only used in one
mode of the Lines table. Pull out logic around table view options into a class,
that handles events on changing view options, and produces an array of records
to pass to Handsontable to render the lines. Remove initial hard-coding of
choice of line records from table configuration, so the only way to get records
for the table is the new LineTableViewOptions class.