Change line "Disable" to "Delete"
Following recommendation from Truss; the mixing of "Disable" and "Delete"
language, and using iconography tied to delete (trash can icon), is confusing.
Simplifying the presentation to always use "Delete" instead of "Disable". The
backend functionality remains the same, if a Line has no Assays or Measurements
linked, it gets truly deleted; otherwise, the "active" flag is set to False to
disable the line and hide it from the normal interface.