Remove dependence on EDDData global variable
Refactored the remaining references to the EDDData global on the Study pages
(lines and data) to instead use methods on the Access class, created in module
scope on getting the initial startup data from the backend.

Change the field definition for Strains to use the findStrain method on Access,
rather than directly accessing the EDDData global variable. Using findStrain is
also shorter than the equivalent direct reference.

Move references to links for study data and for the Study PK to the Access
class, instead of directly loading from the global.

Update the custom event "edddata" to pass the data payload to the event
handler, which eliminates another need to use the global.