Add reverse queryset name for Measurement units
The two foreign keys on Measurement to the X and Y units were configured
without a reverse queryset name. Adding a reverse queryset name allows for
finding which units are used in a given Measurement, Assay, Line, Study, etc.
more easily via the Django ORM. Otherwise, the only way to get this information
would be doing a raw SQL query. The change results in a migration update. The
stand-alone migration is essentially a no-op; it re-defines the foreign key
index, drops the foreign key constraint, then re-adds the constraint with the
same definition. Rather than adding another migration, folding this change into
the existing migration.