Fix resolving of ICE proteins already seen by EDD
The process of resolving an ICE part ID to a protein record in EDD was always
trying to create a new protein record every time. Now there's an additional
check to see if EDD has referenced that ICE part before, and uses the
already-existing protein record from that previous reference.