Fix links to UniProt
We were using an ebi10.uniprot.org subdomain for rendering links to UniProt.
The configuration of UniProt servers has changed, such that the subdomain is
getting redirected to www.uniprot.org instead. In doing this, the server is
misconfigured to omit the initial path separator between the domain and the
path. Switching to using www links should avoid this URL rewrite.