Update setup to use Postgres 11
We're using Postgres 11 now in production, so it only makes sense to update
development environments to use it as well. Eliminating the custom container,
and instead using the official Postgres container. Any existing development
deployments can continue using a compose file with Postgres 9.6; or, update to
use Postgres 11 on a new volume.

Changes to make in compose override file:
-   Switch `postgres` and `ice_db` keys for `image` to `library/postgres:11`
-   Add environment for `POSTGRES_DB=edd` and `POSTGRES_USER=edduser`
-   Change `postgres.env` to use value of `EDD_PGPASS` for `POSTGRES_PASSWORD`