Refactor Study-Lines.ts functions to methods
Encapsulating the two major pieces of functionality in the page into two
classes, one for dealing with the table and controls manipulating the table,
and one for controling the add/edit forms and modals for Lines and Assays.
Doing this refactoring better encapsulates the variables for the table, forms,
and Access facade, so they are no longer global to the file.