Add in_study parameter to REST filters
Using in_study as a parameter in REST calls will allow for using a study ID,
UUID, or slug value, when querying items. This should be a nice short-cut for
clients that are currently querying the study endpoint with the slug, then
parsing out the ID, and in turn passing that along to the API endpoint the
client actually wants.