Fix login failure when Solr is unresponsive
Adding a try-catch and logging a warning whenever the code updating the Solr
indices fails.

Whenever a user logs in, the database record for that user gets updated with
the last login time set to the current time. When the database record updates,
that triggers an update to Solr. But, if Solr is unresponsive or down, the
index update will throw an exception. If that exception is not caught, it
bubbles up until it ultimately fails the login request.

Fixes EDD-1362.