Add API to fetch current user
EDD stores the request object in a threadlocal variable while processing
requests, so it can be available as a "global" variable in contexts where the
request is not explicitly passed as a parameter. This API fetches the user on
the global request, for times when code needs the user, but not necessarily
the request.