Update pytest warnings configs
Always set pytest to emit warnings with "default" strategy, of once per call
site. Ignore warnings from OpenPyXL re: ignoring unsupported extensions used
within an XLSX file.

Note: this has nothing to do with the filename extension. It's more-or-less
plugins used by Excel, that are not supported in the library. Since the library
does not understand the data provided by this "extension", it's warning that
the data for that extension is ignored.