Replace uses of ICE Registry APIs
Deleting the old ICE APIs, and instead using the simplified API from the prior
commit. This is a big reduction in complexity, around a thousand lines of code
removed, and hundreds of statements and branches eliminated. It's also removing
the last of the warnings about hanging sockets after running full tests.