Refactor registration of Celery task serializer
Moving the registration of our custom JSON serialization classes to just prior
to the initialization of the Celery app. Delaying the registration until the
`edd.utilities` module is imported would sometimes result in the serializers
not getting registered in time for use by Celery. The problem goes away if it
becomes impossible to initialize Celery before the registration.