Add tests of appending StudyLog entries
The added code already had close to full coverage, because tests in other
modules were already causing the code to get executed. Adding some explicit
tests for showing that specific actions will result in specific StudyLog
entries added to the database.

In the process of adding tests, also filled in missing functionality from the
first pass at the feature. In particular, the signals for describing a Study
were not getting triggered via experiment description file uploads, nor for
combinatorial line creation UI.

All the other StudyLog entries record a count of the number of lines involved
in the entry (unit operations in the milestone metric terminology). However,
the exports do not operate directly on Study or Line objects, and only directly
reference the MeasurementValue records that get filtered through the export
request filters. To get this information involves two more queries to the
database as part of the exports.

Some small refactorings done with this work:

The REST views module was beginning to get fairly long, with both View/ViewSets
and FilterSets in the same file. Spinning out the FilterSets into their own
filters module, instead.

Also consolidating the GraphQL modules, which made use of the REST filters. It
made no sense to keep one in the "main" module, only to have it subclassed by
the real one in "edd" module.