Refactor TypeScript imports for ES module interop
These changes to the TypeScript configuration and the imports for JQuery and
Dropzone will assist in compatibility with running the code in tests with Jest.
The files remain functionally the same, with JQuery imports simplifying to:

  import "jquery";

The Dropzone imports become a little more unweildy, but those imports are
limited to only a few scripts.