Refactor TypeScript directory structure slightly
Minor re-orgainization of the TypeScript files. The Study-Create.ts file is not
used anywhere, anymore. So it's deleted.

Consolidating typing definition files. The EDDDataInterface.ts file is only
interface declarations; it would get compiled to an EDDDataInterface.d.ts file
that is nearly identical to the original, except for comments and some spacing.
Keeping it as a regular TypeScript file is a waste, so it's moved to the
typings directory and gets a proper .d.ts extension. Moving the Multi-Column
Autocomplete typing definition out of a subfolder and to the root of typings,
since nothing else is there and the directory nesting is unnecessary.