Fix extension issue in import React scripts
The React code for the wizard import to EDD had a small mis-configuration.
Webpack would not build the scripts without having imports retain the `.tsx`
extensions on imports. At the same time, `tsc` and `eslint` and other tools
would complain about the extension existing. So some tools would not work
unless the extension is included, and other tools would not work unless the
extension is omitted. Adding in the `.tsx` extension in the Webpack
resolve.extensions config fixes the issue.