Replace django-rest-swagger with drf-yasg
The project for django-rest-swagger has not seen any updates in over two and a
half years, and has been officially deprecated for eighteen months. It is no
longer compatible with the latest versions of Django, and is busted in our
deployments. The project recommends switching to drf-yasg (for Yet Another
Swagger Generator), so that's what we'll do.