Add parameter for replicate grouping to Lines API
Switching to using paged APIs for fetching data in the front-end causes an
issue with how the front-end is currently computing replicate groupings. All
Lines must be iterated over, to find those that have the same replicate ID.
When only partial Lines get returned to the front-end, it's likely that a
replicate in a grouping can get missed. So, add in a query parameter that does
grouping in the back-end. The grouping is done by annotating the regular query
with a field specifically for replicate ID, then doing paging only on the
unique replicate IDs found. Then, return data where the original query has
replicate IDs in the current page.