Fix log output in some export conditions
In some exports, including worklists, there would be excessive log output
because of inadequate filtering of values before rendering. Specifically, if an
export is selected with a column for strain information, on a Study with one or
more Line objects without strains, then a TypeError would get logged because of
converting a `[None]` to a pipe-delimited string of names. There is a similar
problem in worklists, where the renderer is making use of the `to_json()`
function on `EDDObject`, which tries to render full update objects to JSON; but
the `created`/`updated` fields may not be included in the query
`select_related()`, and the render would fail.

In both these cases, the exports would work fine, there would just be a lot of
log output. The excessive output makes it harder to filter down to
actual problems.