Improve REST in_study parameter
Adding the in_study parameter to more classes of objects in the REST API, and
making a small change to the lookup mechanism by switching the Study lookup to
a subquery. Making that lookup in a subquery can greatly improve performance in
some cases, by fully utilizing the database index. The previous query mechanism
would sometimes result in doing a full sequential scan of joined tables,
because both sides of the join would have restrictions.