Update REST for data used by frontend
This includes:
- Adding an endpoint for measurement compartment;
- Dropping the separate values endpoint, folding into measurements;
- Tweaking serializers to use more client-friendly naming;

With these changes in place, the custom queries that fetch *all* of a Study's
data for use in the frontend can get replaced with using the paged REST API
queries instead. Once the code is further adapted to use paged data, there
should be no more issues with "too much" data locking up the interface.