Add link to REST endpoints in templates
This is a compromise solution, to try to keep the Django and TypeScript
codebases to a minimal number of codependencies. Django renders the link to
REST endpoints as a single URL, that then returns JSON with further REST links.
Changing the links thus needs no updates to the templates. TypeScript has an
interface for the endpoint links it expects.