Update frontend to use paged REST queries
Converts the frontend code for study pages to use the paged REST API queries to
load data, instead of the older `/edddata` pseudo-REST endpoint. Because this
is a wholesale replacement of how data is loaded and manipulated, it's
unfortunately a large commit. There wasn't a natural way for me to break up the
code much further into a set of gradual change. So much of each piece depends
on how all the other pieces work.