Update base image to Bullseye
From previously used Buster base image. The build scripts are modified to
account for this change. The core build now adds a `--pull` flag, as otherwise
the build will continue to use an outdated local copy of the base
`scikit-learn` image.  As a consequence, the build will also attempt to pull an
image for the `node` image running the TypeScript build, so the script for that
now does a build *and* a push.