Update Pipfile dependencies
Several packages have updates from the last re-locking of the Pipfile. This
update is explicitly leaving out the Django 4.0 upgrade for now, as at least
the `graphene-django` package was still incompatible while I was setting up
this commit. The project appears to be close to a release, a new beta version
*is* compatible. Django updates will wait until then, and further testing.

Resolves patch alerts for `celery` and `python-ldap`.