Remove outdated AJAX calls
These removed calls are no longer used, replaced with the `/access/` URL giving
links to the REST API to load in data. A few of the outdated calls still exist
here, because they are still in use by the legacy imports--from the `edd.load`
package and the cytometry import in utilities. Once these are removed, the
remainder of the EDDData and ATData code can be removed.