Remove flower container and code
There's been dependabot alerts for multiple items in the flower image for a
little while. We're not using it to manage task queues, and AFAIK, it was never
really working correctly in the first place. Rather than expending effort on
keeping something we're not using up-to-date, I'm just deleting it. Should it
become important later, this can always be reverted and updated to meet our
actual needs.