Change default y-axis bound calculation
Requested by Chris, the default calculation of setting the axis to minimum and
maximum displayed values makes for a misleading display, especially when only
showing two items. The default will now constrain miminum to zero if the
minimum is greater than zero. Follow up work for EDD-1418 should allow for
users to change the range used for each axis.