Remove old CarbonSource table/metadata
We've decided to purge the CarbonSource table from EDD. The setup of the table
was based on the very early conceptions of EDD, and the functionality provided
by it did not match what we actually wanted EDD to do. So, all existing
CarbonSource records are getting merged into the metadata field for "Media" as
free-text, and the table is getting dropped.

Most of the existing references to CarbonSource are already removed from the
code, but there are a few still lingering around. These are also removed along
with the table, completing the deprecation.