Fix display of paging controls for readonly users
The new paging controls on the Lines/Description view were only being added to
EDD when the current user has write-access to a Study. This oversight makes EDD
substantially less useful to readonly users. Whoooops!

Checking for any similar problems on the Data view, the Import button on that
view was always getting displayed, even if the current user does not have write
access to the Study. This should now also be fixed.