Fix describe templatetag tests
The existing tests were asserting assumptions that are not relevant to the
functionality under test. Specifically, the tests were assuming a specific
static files setup. Any environment that does not serve static assets from
`/static` would thus fail these tests, even if the functionality worked
correctly. The tests are now updated to assert that the example file names
are rendered in the simple templates used to test.