Add Line replicate key to REST exports
Since EDD has support for tracking Line replicate groups now, the exports
should also support including the replicate grouping keys in its output. There
is now an additional column in exported CSVs for "Replicate Key", set to the
Line UUID if no replicate grouping is set, and set to the shared replicate key
value for Lines that are in a replicate grouping. This ensures that every Line
has some value in the column, that can then be used to group/filter the CSV
data in downstream tools.