Fix instances of assumed settings values
This fix is the result of scanning the code for appearances of code like
`settings.MY_SETTING`, and replacing with
`getattr(settings, "MY_SETTING", DEFAULT)`. The only remaining appearances of
the former pattern are with Django settings that are known to have a default,
`settings.AUTH_USER_MODEL` and `settings.SERVER_EMAIL`. These are left as-is to
prevent causing migrations, etc.

Taking care of this now, because one of these invalid access patterns is
causing a bug currently in viewing the admin pages for Protein records.

Closes: EDD-1218