Add guard for non-string keys in settings filter
I found it was possible for the filter responsible for sanitizing passwords in
URLs in the Django settings to get a non-string (probably `None`) key value.
This results in an exception, in code that is handling a different unhandled
exception, and generally makes a mess of everything. Adding in a test for the
key being a string before testing it on the regular expression that looks for
URLs to filter.