Fix resolution of protein names from UniProt
Two failure modes around protein creation are fixed here. First, the function
fetching data from UniProt was returning a dict with the protein data, but was
used in a context where an iterator of key-value pairs is required. Returning
the `items()` iterator instead of the dict resolves the problem. Second, the
picking of the first valid name was allowing empty or whitespace strings to
pass. Doing this results in name-less proteins, when we want to use the
fallback value instead. The resolution is now properly checking if the name
would be "blank" and using the fallback value.