Remove unused environment from TypeScript build
The `EDD_VERSION` variable is added to the build phase for the TypeScript code,
but is not used. This only serves to prevent using the cache, even if the base
Node image and the TypeScript code is not changing. Removing this should cut
out around 90s for builds that have no changes to TypeScript code.