Fix error handling with "Omit Strains" option
When a value in an Experiment Description file "Strain" column cannot be found
in the linked ICE registry, an error with a button to "Omit Strains" is shown.
Clicking the button is supposed to silence the validation errors for the
strains not found, and proceed as if there was no value entered at all. Because
the error handling logic of this part of the code is so convoluted, this option
to skip errors got dropped when updating to use the improved ICE querying APIs.

Fixes: EDD-1399