Add detection of logout to websockets
This has bugged me for a while. When a user has an EDD page open in a tab for a
long time, and their normal HTTP session expires, the websocket connection
stays open. This is not detectable normally, but when new versions are
deployed, the server closes old websockets. The frontend code tries to
reconnect, and fails--because the session is now invalid. That open tab will
then continue trying to reconnect, until it is closed or the user tries to
navigate elsewhere.

With this detection code added, when the reconnect attempt fails a few times,
it will kick off a "ping" request to EDD, which will fail if the current
session is invalid. If the session does turn out to be invalid, the code will
reload the page, to force the user to login again. This process has to be
outside of the websocket connection, because the protocol intentionally does
not provide any information on why connections fail.

Fixes: EDD-1333