Add show password button to login page
Having a password field always obscure the field contents can make the field
difficult to use if the user is unclear of what they have typed (e.g. has
tremors, types fast). Having the option to reveal the contents of the password
field can allow a user to review it, and correct it, without needing to submit
the form and fill out the password all over again.

This feature is behind a feature flag, and is disabled by default. There is a
trade-off between usability and security for a feature like this. Since the
LBNL deployments of EDD accept LBNL LDAP credentials to login, it is an
inappropriate accommodation to make considering the consequences of
unintentional password reveal.