Add archive of coverage.json to builds
As Jenkins is running through tests anyway, it's convenient to archive the
coverage information as a downloadable file. This will allow for current
coverage information on a branch to be available for any branch built by
Jenkins, without having to locally re-run the tests to generate the file. This
can save several minutes in getting the coverage information to editors if,
e.g. several changes hit `trunk` at once. Just download the file, drop in
place, and current coverage is available.