Move node build into core Dockerfile
Having a unified Dockerfile allows for eliminating the standalone node
directory, and eliminating the argument passed for the node version to the main
build. It's still possible to build an independent node image for doing
development, and the script is updated to do so. This just simplifies the
overall build.