Fix W3C validation issues on Study Overview
A number of issues in HTML validation are on the Study Overview page, as
identified by the Nu HTML Checker at https://validator.w3.org/nu/.

As a side-effect, two items in page scripts need to update. First, updating
error message template elements requires updating the selectors used to
populate the template elements in the DescriptionDropzone.ts file. Second, the
switch from a non-standard attribute for autocomplete metadata, to a standard
data-* attribute, requires switching the autocomplete APIs to use a JQuery
data() call instead of an attr() call.