Update Solr healthcheck to match startup script
Or, match as closely as possible to the checks used in the startup script in a
bash one-liner without requiring any other programs to be installed to do
proper JSON parsing.

This update brings the internal healthcheck more in line with the check used by
EDD to see if the indexing service is up. We found there were problems when
restoring Solr after (I think?) a Docker upgrade. The healthcheck Docker ran on
the /admin/info/health endpoint was returning OK, while the rest of Solr was
not responding. The EDD checks were instead using /admin/info/system, and
getting not-OK responses. Or timing out. I don't recall specifics anymore.

Fixes: EDD-1336