Update Django i18n templatetag names
The canonical names for the Django translation templatetags are updated,
switching from `trans` to `translate`, and `blocktrans` / `endblocktrans` to
`blocktranslate` / `endblocktranslate`. The prior names will still function,
but are no longer documented. Updating our uses of these templatetags to use
the un-abbreviated and more explicit new canonical names.

Did this with two commands:

    find . -type f -path '*/templates/*' -exec sed -i "" -E 's/blocktrans /blocktranslate /g' {} \;
    find . -type f -path '*/templates/*' -exec sed -i "" -E 's/{% trans /{% translate /g' {} \;

Fixes: EDD-1431