Update dev mode refresh_static.sh script
The previous configuration for refresh_static.sh was making some assumptions
that were particular to my MacOS setup at the time. All my dev is on Linux now,
and those assumptions were made apparent, and shown to be wrong. The script is
now using normal application settings, and the previous settings--based on a
variation of the Dockerfile build settings--are removed.