Add mkdocs improvements
Including some improvements to MkDocs for generating the EDD documentation
site, that are in use in some other ESE static documentation sites. Adds the
`admonition` extension, so the docs can include callout boxes like "Info",
"Warning", etc. Adds an override to the default navigation depth on the theme,
so the TOC generated can go further than just the top two levels of headers.