Fix migration detection
The renaming of the Protocol model in 1183c6bb1f0 is causing the Django
migrations code a bit of trouble. The existing migrations are successfully
updating the database state, but Django is getting stuck in an infinite loop
when detecting migrations. The fix seems to be removing the offending fields
entirely. The fields on Line and Study are not used anywhere, and are safe to
permanently remove; the same information is available by directly querying the
Protocol model.

The field on Category in the load app is also stuck, but is in use. Removing it
requires switching the Import Wizard in the `load` module to use an
autocomplete for filtering Protocols, rather than using a list configured via
the admin interface.