Update base image installed package versions
Add another Dockerfile build stage, replacing the reference to the scikit-learn
base image. It turns out, none of the code in EDD references scipy or
scikit-learn, so we should not need to install them, and can save roughly 200MB
of space.

Bumping the Python version to 3.10, and removing the restriction preventing
Django from upgrading to 4.0. Forcing `graphene-django` to `3.15.0`, as it
technically does not claim Django 4.0 support and would otherwise install
version `3.0.0`. Include empty `*.map` files so the static files in `drf-yasg`
can properly get collected with references to content-hashed filenames.