Consolidate Django template HTML/text
The locations for the various templates used in the project could be difficult
to determine. Since EDD arranges its code in multiple "apps", and each one
could potentially include a templates folder, finding the location of a
template file referenced in the code would involve checking all of them--or
running `find . -name page.html` and picking out the correct one if multiples
are found. Bundling templates this way is more important for re-usable "apps",
and none of our current code is written to be used that way. These are only for
EDD use.

This update consolidates all of our customized templates into a high-level
templates directory. Inside are all the templates for our internal apps, and
the override templates for other apps (e.g. `account` for the `django-allauth`
library app). This prevents any potential confusion about where templates
should go. They all go to the templates folder. It also eliminates problems
where a template could potentially exist in multiple locations.